LearningByAI/mockplus/Gemini/login-app/backend/requirements.txt
# backend/requirements.txt
Flask
Flask-Cors
Pillow
